backend/requirements.txt
# --- FastAPI 기본 ---
fastapi==0.118.0
uvicorn==0.37.0
starlette==0.48.0
python-dotenv==1.1.1

# --- 인증 / 보안 ---
PyJWT==2.10.1
bcrypt==5.0.0
passlib==1.7.4

# --- 데이터베이스 (SQLite 기반) ---
sqlalchemy==2.0.36
pydantic==2.11.9
pydantic_core==2.33.2

# --- 유틸리티 ---
requests==2.32.4
python-dateutil==2.8.2
typing_extensions==4.12.2

# --- Notion API (선택적으로 유지) ---
notion-client==2.5.0
